Create Order Service Implementation
Create Order Service Implementation

+++ service-api/src/main/java/com/d13k_mbp/service_api/dto/response/ResponseOrderItemDTO.java
@@ -18,6 +18,6 @@ public class ResponseOrderItemDTO {
     private BigDecimal pricePerUnit;
     private int itemDiscount;
     private BigInteger itemQuantity;
-    private BigInteger totalPrice;
+    private BigDecimal totalPrice;
     private LocalDateTime createdAt;
 }
